requirements.txt
attrs==23.1.0
azure-core==1.29.4
azure-eventhub==5.11.4
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
jsonlines==4.0.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.6
